requirements.txt
astroid==2.3.2
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.4.3
rope==0.14.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
azure-functions
